Add error handler for all spawned processes
Add .on('error') handler for all spawned processes.
They are required when the process cannot be spawned.
As the spawn is not sync, if the process cannot be created, the error is raised on process.nextTick.
If 'error' handler is missing, the error that is thrown on process.nextTick cannot be catched and the process crashes.
When error handler is added, the code can be try-catched.

In order to reproduce the issue, try using winreg on Linux or Mac.